redundant cast removed

+++ src/vita/core/conv.c
@@ -13,7 +13,7 @@ int64_t conv_str_to_int64(const char *const z) {
     }
 
     // save value
-    return (int64_t)strtoll(z, NULL, 10);
+    return strtoll(z, NULL, 10);
 }
 
 uint64_t conv_str_to_uint64(const char *const z) {
@@ -29,7 +29,7 @@ uint64_t conv_str_to_uint64(const char *const z) {
     }
 
     // save value
-    return (uint64_t)strtoull(z, NULL, 10);
+    return strtoull(z, NULL, 10);
 }
 
 double conv_str_to_double(const char *const z) {
@@ -45,7 +45,7 @@ double conv_str_to_double(const char *const z) {
     }
 
     // save value
-    return (double)strtod(z, NULL);
+    return strtod(z, NULL);
 }
